pyproject.toml
[tool.poetry]
name = "docuhive"
version = "0.1.0"
description = "This application can scan through documents to extract information"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11.0 || >3.11.0,<4.0"
opencv-python-headless = "^4.7.0.72"
google-cloud-vision = "^3.4.1"
pillow = "^9.4.0"
pymupdf = "^1.21.1"
celery = "^5.2.7"
flask = "^2.2.3"
sqlalchemy = "^2.0.7"
strawberry-graphql = "^0.168.1"
numpy = "^1.24.2"
pytesseract = "^0.3.10"
flask-cors = "^3.0.10"
flask-sqlalchemy = "*"
graphviz = "^0.20.1"
pandas = "^2.0.0"
pdf2image = "^1.16.3"
sentence-transformers = "^2.2.2"
scikit-learn = "^1.2.2"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
gevent = "^22.10.2"
psycopg2 = "^2.9.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
black = "^23.3.0"
diff-cover = "^7.5.0"
pytest-cov = "^4.0.0"
scalene = "^1.5.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
